fix: Update HeatmapData and HeatMapView to use Int instead of Double

+++ EyeGazeDisplay/EyeGazeDisplay/src/View/HeatMapView.swift
@@ -69,8 +69,8 @@ struct HeatMapView: View {
 
     }
     
-    func colorForValue(_ value: Double) -> Color {
-        let normalizedValue = value / 100.0 //正規化
+    func colorForValue(_ value: Int) -> Color {
+        let normalizedValue = Double(value) / 100.0 //正規化
         let red = 0.25 + (0.75 * normalizedValue) // 0.25から1.0の範囲で赤を増加させる
         let green = 0.2 * (1.0 - normalizedValue) // 0.2から0.0の範囲で緑を減少させる
         return Color(red: red, green: green, blue: 0.0, opacity: normalizedValue)
